docker/instrumented-app/requirements.txt
flask==3.0.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-flask==0.42b0
opentelemetry-exporter-otlp==1.21.0
opentelemetry-exporter-prometheus==0.42b0
prometheus-client==0.19.0
requests==2.31.0

